Fix a performance issue in TTD's TraceAnalysis sample (#131)
It uses a hot-path callback that pushes data to a vector.

`std::vector` is designed such that reallocations are infrequent, but even these infrequent reallocations can hurt concurrency greatly.

In this case, a simple improvement is to merge the elements whenever the vector reaches its capacity. Merging reduces the vector size, creating free capacity without interacting with the global heap.